requirements.txt
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
idna==3.8
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pybit==5.8.0
pycryptodome==3.20.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
seaborn==0.13.2
setuptools==74.0.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
